테스트 코드 수정 및 extension 코드 다른 Swift 파일로 이동
