+++ FoodNutrition/RestaurantMapView.swift
@@ -10,17 +10,18 @@ import MapKit
 
 struct RestaurantMapView: View {
 	
+	let restaurantName: String
 	let restaurantCoordinate: CLLocationCoordinate2D
 	
 	var body: some View {
 		Map(initialPosition: MapCameraPosition.region(MKCoordinateRegion(center: restaurantCoordinate, latitudinalMeters: 400, longitudinalMeters: 400))) {
-			Marker("강강술래 늘봄농원점", coordinate: restaurantCoordinate)
+			Marker(restaurantName, coordinate: restaurantCoordinate)
 		}
 	}
 }
 
 #Preview {
-	RestaurantMapView(restaurantCoordinate: CLLocationCoordinate2D(latitude: 37.6877324, longitude: 126.9013021))
+	RestaurantMapView(restaurantName: "강강술래 늘봄농원점", restaurantCoordinate: CLLocationCoordinate2D(latitude: 37.6877324, longitude: 126.9013021))
 		.padding()
 		.clipShape(.rect(cornerRadius: 50))
 		.frame(height: 200)
